backend/email-parser/requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
google-generativeai
